+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -10,6 +10,9 @@ import java.util.ArrayList;
 import java.util.List;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.servlet.FrontServlet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
