--- tomcat/src/main/java/org/apache/coyote/http11/controller/LoginController.java
@@ -1,18 +1,0 @@
-package org.apache.coyote.http11.controller;
-
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
-import org.apache.coyote.http11.HttpStatus;
-
-public class LoginController extends Controller {
-    @Override
-    public void processPost(final HttpRequest httpRequest, final HttpResponse httpResponse) {
-
-    }
-
-    @Override
-    public void processGet(final HttpRequest httpRequest, final HttpResponse httpResponse) {
-        httpResponse.setStatus(HttpStatus.OK);
-        httpResponse.setView("login");
-    }
-}
