+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpHeaders.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
