+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestLine.java
@@ -1,9 +1,10 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
+import org.apache.coyote.http11.response.HttpMethod;
 
 public class RequestLine {
