+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/HttpRequestTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.RequestLine;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.request.RequestLine;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
