+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/RequestLineTest.java
@@ -3,8 +3,8 @@ package nextstep.org.apache.coyote.http11;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import org.apache.coyote.http11.HttpMethod;
-import org.apache.coyote.http11.RequestLine;
+import org.apache.coyote.http11.request.RequestLine;
+import org.apache.coyote.http11.response.HttpMethod;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
