+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/StatusLineTest.java
@@ -2,8 +2,8 @@ package nextstep.org.apache.coyote.http11;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.apache.coyote.http11.HttpStatus;
-import org.apache.coyote.http11.StatusLine;
+import org.apache.coyote.http11.response.HttpStatus;
+import org.apache.coyote.http11.response.StatusLine;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
